+++ cube_sim.py
@@ -27,50 +27,42 @@ class cube:
         self.actionList = [self.clockwise, self.counterclockwise,
                             self.forward, self.backward]
 
+        self.al = [cube.clockwise,cube.counterclockwise,cube.forward,cube.backward]
+
+        self.faces = {'front':self.front,'top':self.top,'bottom':self.bottom,
+                                'left':self.left,'right':self.right,'back':self.back}
     
     def update_cubeState(self):
         self.cubeState = np.array([self.front,self.top,self.bottom,
                                 self.left,self.right,self.back]
                                 ).reshape(6,self.size,self.size)
-    
-    def getActions(self):
-        feat = self.get_features()
-        if feat[-1] == 1:
-            return list()
-        return self.actionList
+        self.faces = {'front':self.front,'top':self.top,'bottom':self.bottom,
+                                'left':self.left,'right':self.right,'back':self.back}
 
     #===============================================
     # Get Features
     #===============================================
 
     def get_features(self):
-        feat_vector = []
+        feat_vector = dict()
         tot_cpf = 0
         one_color_face = 0
-        for face in self.cubeState:
+        for face in self.faces:
             # calc colors per face
-            num_col = len(np.unique(face))
+            num_col = len(np.unique(self.faces[face]))
             tot_cpf += num_col
             # Add 1/color for that face to feature vector
-            feat_vector.append(1/num_col)
+            feat_vector[f'inv_cpf_{face}'] = (1/num_col)
             # get total single colored faces
             if num_col == 1:
                 one_color_face += 1
         # Add 1 / average colors per face
-        feat_vector.append(1/(tot_cpf/6))
+        # feat_vector['inv_avg_cpf'] = (1/(tot_cpf/6))
         # Add number of single colored faces (divided by 6 to keep same scale as other features)
-        feat_vector.append(one_color_face/6)
+        feat_vector['one_color_faces'] = (one_color_face/6)
         return feat_vector
 
     def copy(self):
-        # cubeCopy = cube()
-        # cubeCopy.front = self.front
-        # cubeCopy.top = self.top
-        # cubeCopy.bottom = self.bottom
-        # cubeCopy.left = self.left
-        # cubeCopy.right = self.right
-        # cubeCopy.back = self.back
-        # return copy.deepcopy(cubeCopy)
         return copy.deepcopy(self)
 
     def getFace(self,cubeState,face):
